refactor(dispatch/router): move router tests to separate tests module
Extract the large in-file test module from routers.rs into its own dedicated file tests.rs within the dispatch/router directory. This improves code organization and test maintainability without changing test logic or functionality.
